Add peltool
This is a python based PEL parser that outputs a PEL as JSON.  It uses
the same SRC and UserData plugin infrastructure as the original C++
based peltool.

It uses the 'pel_registry' module to find where the
message_registry.json and component name files are installed, though if
that isn't present, which it won't be on the BMC, it will look in the
BMC location which is /usr/share/phosphor-logging/pels/.  If it still
can't find it, it will just skip the SRC details and component names.

Invoke with:
$ python3 -m pel.peltool.peltool -f <PEL file>
$ python3 .../site-packages/pel/peltool/peltool.py -f <PEL file>

For now, it isn't installing peltool.py as a script into bin, since I
don't think I'm ready for all BMC users to know about it yet, which
they would if it was in the path along with the original peltool.

Signed-off-by: George Liu <liuxiwei@inspur.com>
Signed-off-by: Matt Spinler <spinler@us.ibm.com>