peltool: (non)serviceable PEL skipping
Add options to peltool so it will just exit without printing anything
based on if the PEL is serviceable or not, where a serviceable PEL is
one with a severity that is neither informational or recovered.

This is for the case where a script is calling peltool on many PELs to
allow it to choose if it wants all PELs printed, or just serviceable or
nonserviceable ones.

-s: only print PEL if it's serviceable
-n: only print PEL if it's nonserviceable

Signed-off-by: Matt Spinler <spinler@us.ibm.com>