peltool: Use procedure desc plugins
The maintenance procedures used in callouts may have a description
provided by a plugin, so attempt to load the plugin and get the
description.

The plugin would be named calloutparsers.Xcallouts.Xcallouts where X is
the creator ID of the PEL in lower case.  The function called is:

    def getMaintProcDesc(procedure: str) -> str

It returns a JSON string.

Signed-off-by: Matt Spinler <spinler@us.ibm.com>