Fix hexdump for partial lines
After the previous commit that changed the default format of a hex dump,
the math was wrong when the number of characters per line was calculated
as the spacing between chunks changed to 2 from 1.  This only showed up
when there was less than 16 bytes per line.

The testcases are all still out of date, they'll be fixed in a separate
commit.

Signed-off-by: Matt Spinler <spinler@us.ibm.com>